==> 2_MWAS_AdditionalAnalyses/3_Report_Part4_CaseControl_sex-specific.Rmd <==
---
title: "Report analysis - PART 4"
output:
  html_document:
    df_print: paged
  pdf_document: default
header-includes: "\\usepackage[export]{adjustbox}\n\\let\\includegraphicsbak\\includegraphics\n\\renewcommand*{\\includegraphics}[2][]{\\includegraphicsbak[frame,#1]{#2}}
  \n"
---

```{r}
# LOAD PATHS
source("S:/Project/WP-epigenetics/04_Pipeline_PhaseI/2_Paths.R")
dir_CaseCtrl <- "S:/Project/WP-epigenetics/10_CaseCtrl_EWAS/"
```


### -> sex-specific analysis

# FEMALE SAMPLE


```{r include=FALSE}

library(readxl)

samplesheet <- as.data.frame(read_xlsx(file.path(dir_CaseCtrl,"output_F/Tables/Final_samplesheet_CaseCtrl.xlsx")))

```


## FINAL SAMPLESHET OVERVIEW

### Sex
```{r, echo=FALSE, message=FALSE}
print(table(samplesheet$Sex))
```
### Case-Control status
```{r, echo=FALSE, message=FALSE}
print(table(samplesheet$Case_Control))
```

### Age

```{r, echo=FALSE, message=FALSE}
print(paste0("Minimum age: ", min(samplesheet$Age)))
print(paste0("Maximum age: ", max(samplesheet$Age)))
print(paste0("Median age: ", median(samplesheet$Age)))
print(paste0("Mean age: ", mean(samplesheet$Age)))
print(paste0("sd age: ", sd(samplesheet$Age)))
print(paste0("IQR age: ", IQR(samplesheet$Age)))
```

### Ancestry
```{r, echo=FALSE, message=FALSE}
print(table(samplesheet$Ancestry))
```


## QQ plots

### 2 ctrl probe PCs, 2 ancestry PCs
```{r, out.width="400px", out.height="400px", message=FALSE, echo=FALSE}
knitr::include_graphics(file.path(dir_CaseCtrl,"output_F/Figures/QQplots/QQplot_CaseCtrl_Females_2CtrlPCs_2AncPCs_limma.png"))
```

### 5 ctrl probe PCs, 2 ancestry PCs
```{r, out.width="400px", out.height="400px", message=FALSE, echo=FALSE}
knitr::include_graphics(file.path(dir_CaseCtrl,"output_F/Figures/QQplots/QQplot_CaseCtrl_Females_5CtrlPCs_2AncPCs_limma.png"))
```

### 10 ctrl probe PCs, 2 ancestry PCs
```{r, out.width="400px", out.height="400px", message=FALSE, echo=FALSE}
knitr::include_graphics(file.path(dir_CaseCtrl,"output_F/Figures/QQplots/QQplot_CaseCtrl_Females_10CtrlPCs_2AncPCs_limma.png"))
```

### 15 ctrl probe PCs, 2 ancestry PCs
```{r, out.width="400px", out.height="400px", message=FALSE, echo=FALSE}
knitr::include_graphics(file.path(dir_CaseCtrl,"output_F/Figures/QQplots/QQplot_CaseCtrl_Females_15CtrlPCs_2AncPCs_limma.png"))
```

### 2 ctrl probe PCs, 5 ancestry PCs
```{r, out.width="400px", out.height="400px", message=FALSE, echo=FALSE}
knitr::include_graphics(file.path(dir_CaseCtrl,"output_F/Figures/QQplots/QQplot_CaseCtrl_Females_2CtrlPCs_5AncPCs_limma.png"))
```

### 5 ctrl probe PCs, 5 ancestry PCs
```{r, out.width="400px", out.height="400px", message=FALSE, echo=FALSE}
knitr::include_graphics(file.path(dir_CaseCtrl,"output_F/Figures/QQplots/QQplot_CaseCtrl_Females_5CtrlPCs_5AncPCs_limma.png"))
```

### 10 ctrl probe PCs, 5 ancestry PCs
```{r, out.width="400px", out.height="400px", message=FALSE, echo=FALSE}
knitr::include_graphics(file.path(dir_CaseCtrl,"output_F/Figures/QQplots/QQplot_CaseCtrl_Females_10CtrlPCs_5AncPCs_limma.png"))
```

### 15 ctrl probe PCs, 5 ancestry PCs
```{r, out.width="400px", out.height="400px", message=FALSE, echo=FALSE}
knitr::include_graphics(file.path(dir_CaseCtrl,"output_F/Figures/QQplots/QQplot_CaseCtrl_Females_15CtrlPCs_5AncPCs_limma.png"))
```

### 2 ctrl probe PCs, 10 ancestry PCs
```{r, out.width="400px", out.height="400px", message=FALSE, echo=FALSE}
knitr::include_graphics(file.path(dir_CaseCtrl,"output_F/Figures/QQplots/QQplot_CaseCtrl_Females_2CtrlPCs_10AncPCs_limma.png"))
```

### 5 ctrl probe PCs, 10 ancestry PCs
```{r, out.width="400px", out.height="400px", message=FALSE, echo=FALSE}
knitr::include_graphics(file.path(dir_CaseCtrl,"output_F/Figures/QQplots/QQplot_CaseCtrl_Females_5CtrlPCs_10AncPCs_limma.png"))
```

### 10 ctrl probe PCs, 10 ancestry PCs
```{r, out.width="400px", out.height="400px", message=FALSE, echo=FALSE}
knitr::include_graphics(file.path(dir_CaseCtrl,"output_F/Figures/QQplots/QQplot_CaseCtrl_Females_10CtrlPCs_10AncPCs_limma.png"))
```

### 15 ctrl probe PCs, 10 ancestry PCs
```{r, out.width="400px", out.height="400px", message=FALSE, echo=FALSE}
knitr::include_graphics(file.path(dir_CaseCtrl,"output_F/Figures/QQplots/QQplot_CaseCtrl_Females_15CtrlPCs_10AncPCs_limma.png"))
```

## LAMBDA VALUES

```{r, message=FALSE, echo=FALSE}
lambda <- readxl::read_xlsx(file.path(dir_CaseCtrl, "output_F/Tables/Lambda_CaseCtrl_Females_limma.xlsx"))
knitr::kable(lambda)
```

## VIF (TEST FOR MULTI-COLINEARITY)

```{r, message=FALSE, echo=FALSE}
vif <- readxl::read_xlsx(file.path(dir_CaseCtrl, "output_F/Tables/VIF_CaseCtrl_Females.xlsx"))
knitr::kable(vif)
```






# MALE SAMPLE



```{r include=FALSE}

library(readxl)

samplesheet <- as.data.frame(read_xlsx(file.path(dir_CaseCtrl,"output_M/Tables/Final_samplesheet_CaseCtrl.xlsx")))

```


## FINAL SAMPLESHET OVERVIEW

### Sex
```{r, echo=FALSE, message=FALSE}
print(table(samplesheet$Sex))
```
### Case-Control status
```{r, echo=FALSE, message=FALSE}
print(table(samplesheet$Case_Control))
```

### Age

```{r, echo=FALSE, message=FALSE}
print(paste0("Minimum age: ", min(samplesheet$Age)))
print(paste0("Maximum age: ", max(samplesheet$Age)))
print(paste0("Median age: ", median(samplesheet$Age)))
print(paste0("Mean age: ", mean(samplesheet$Age)))
print(paste0("sd age: ", sd(samplesheet$Age)))
print(paste0("IQR age: ", IQR(samplesheet$Age)))
```

### Ancestry
```{r, echo=FALSE, message=FALSE}
print(table(samplesheet$Ancestry))
```


## QQ plots

### 2 ctrl probe PCs, 2 ancestry PCs
```{r, out.width="400px", out.height="400px", message=FALSE, echo=FALSE}
knitr::include_graphics(file.path(dir_CaseCtrl,"output_M/Figures/QQplots/QQplot_CaseCtrl_Males_2CtrlPCs_2AncPCs_limma.png"))
```

### 5 ctrl probe PCs, 2 ancestry PCs
```{r, out.width="400px", out.height="400px", message=FALSE, echo=FALSE}
knitr::include_graphics(file.path(dir_CaseCtrl,"output_M/Figures/QQplots/QQplot_CaseCtrl_Males_5CtrlPCs_2AncPCs_limma.png"))
```

### 10 ctrl probe PCs, 2 ancestry PCs
```{r, out.width="400px", out.height="400px", message=FALSE, echo=FALSE}
knitr::include_graphics(file.path(dir_CaseCtrl,"output_M/Figures/QQplots/QQplot_CaseCtrl_Males_10CtrlPCs_2AncPCs_limma.png"))
```

### 15 ctrl probe PCs, 2 ancestry PCs
```{r, out.width="400px", out.height="400px", message=FALSE, echo=FALSE}
knitr::include_graphics(file.path(dir_CaseCtrl,"output_M/Figures/QQplots/QQplot_CaseCtrl_Males_15CtrlPCs_2AncPCs_limma.png"))
```

### 2 ctrl probe PCs, 5 ancestry PCs
```{r, out.width="400px", out.height="400px", message=FALSE, echo=FALSE}
knitr::include_graphics(file.path(dir_CaseCtrl,"output_M/Figures/QQplots/QQplot_CaseCtrl_Males_2CtrlPCs_5AncPCs_limma.png"))
```

### 5 ctrl probe PCs, 5 ancestry PCs
```{r, out.width="400px", out.height="400px", message=FALSE, echo=FALSE}
knitr::include_graphics(file.path(dir_CaseCtrl,"output_M/Figures/QQplots/QQplot_CaseCtrl_Males_5CtrlPCs_5AncPCs_limma.png"))
```

### 10 ctrl probe PCs, 5 ancestry PCs
```{r, out.width="400px", out.height="400px", message=FALSE, echo=FALSE}
knitr::include_graphics(file.path(dir_CaseCtrl,"output_M/Figures/QQplots/QQplot_CaseCtrl_Males_10CtrlPCs_5AncPCs_limma.png"))
```

### 15 ctrl probe PCs, 5 ancestry PCs
```{r, out.width="400px", out.height="400px", message=FALSE, echo=FALSE}
knitr::include_graphics(file.path(dir_CaseCtrl,"output_M/Figures/QQplots/QQplot_CaseCtrl_Males_15CtrlPCs_5AncPCs_limma.png"))
```

### 2 ctrl probe PCs, 10 ancestry PCs
```{r, out.width="400px", out.height="400px", message=FALSE, echo=FALSE}
knitr::include_graphics(file.path(dir_CaseCtrl,"output_M/Figures/QQplots/QQplot_CaseCtrl_Males_2CtrlPCs_10AncPCs_limma.png"))
```

### 5 ctrl probe PCs, 10 ancestry PCs
```{r, out.width="400px", out.height="400px", message=FALSE, echo=FALSE}
knitr::include_graphics(file.path(dir_CaseCtrl,"output_M/Figures/QQplots/QQplot_CaseCtrl_Males_5CtrlPCs_10AncPCs_limma.png"))
```

### 10 ctrl probe PCs, 10 ancestry PCs
```{r, out.width="400px", out.height="400px", message=FALSE, echo=FALSE}
knitr::include_graphics(file.path(dir_CaseCtrl,"output_M/Figures/QQplots/QQplot_CaseCtrl_Males_10CtrlPCs_10AncPCs_limma.png"))
```

### 15 ctrl probe PCs, 10 ancestry PCs
```{r, out.width="400px", out.height="400px", message=FALSE, echo=FALSE}
knitr::include_graphics(file.path(dir_CaseCtrl,"output_M/Figures/QQplots/QQplot_CaseCtrl_Males_15CtrlPCs_10AncPCs_limma.png"))
```

## LAMBDA VALUES

```{r, message=FALSE, echo=FALSE}
lambda <- readxl::read_xlsx(file.path(dir_CaseCtrl, "output_M/Tables/Lambda_CaseCtrl_Males_limma.xlsx"))
knitr::kable(lambda)
```

## VIF (TEST FOR MULTI-COLINEARITY)

```{r, message=FALSE, echo=FALSE}
vif <- readxl::read_xlsx(file.path(dir_CaseCtrl, "output_M/Tables/VIF_CaseCtrl_Males.xlsx"))
knitr::kable(vif)
```
